Simplify deploy clojure tools.
diff --git a/bootstrap-install-system.sh b/bootstrap-install-system.sh
new file mode 100755
index 0000000..425e27e
--- /dev/null
+++ b/bootstrap-install-system.sh
@@ -0,0 +1,15 @@
+#!/bin/bash -xe
+
+# This is a bootstrap script to install the needed script
+# for installing my overall system.
+# To deploy this script, download it directly from github and make it executable.
+# wget https://github.com/ardumont/sh/bootstrap-install-system.sh
+# chmod +x bootstrap-install-system.sh
+# then execute it.
+
+sudo aptitude install -y git tig emacs24 tmux ruby gitk
+
+# you must have pre-setup your ssh key to access the github account
+CLONE_DIR=$HOME/repo/perso
+
+[ ! d $CLONE_DIR/sh ] && git clone git@github.com:ardumont/sh.git $CLONE_DIR/sh && ln -s $CLONE_DIR/sh $HOME/bin

+++ deploy/deploy-tools-clojure2.sh
@@ -1,15 +1,6 @@
 #!/bin/bash
 
 cd ~/bin
-wget -O lein2 - https://raw.github.com/technomancy/leiningen/preview/bin/lein
-chmod +x ~/bin/lein2
+[ ! -f ~/bin/lein ] && wget -O lein2 - https://raw.github.com/technomancy/leiningen/preview/bin/lein && chmod +x ~/bin/lein2 && ln -nfs ~/bin/lein2 ~/bin/lein
 
-lein2 -v
-
-# install the plugin to make a server swank
-lein2 plugin install swank-clojure 1.4.2
-lein2 plugin install lein-midje 1.0.9
-lein2 plugin install lein-marginalia 0.7.0
-lein2 plugin install lein-noir 1.2.1
-lein2 plugin install lein-cljs 0.2.2
-
+lein -v
